Refactor objects
