+++ src/openrct2-ui/windows/Footpath.cpp
@@ -24,7 +24,7 @@
 #include <openrct2/interface/viewport.h>
 #include <openrct2/interface/widget.h>
 #include <openrct2/localisation/localisation.h>
-#include <openrct2/ObjectList.h>
+#include <openrct2/object/ObjectLimits.h>
 #include <openrct2/sprites.h>
 #include <openrct2/world/footpath.h>
 #include <openrct2-ui/interface/Dropdown.h>
