+++ src/openrct2-ui/windows/Scenery.cpp
@@ -24,7 +24,7 @@
 #include <openrct2/interface/widget.h>
 #include <openrct2/localisation/localisation.h>
 #include <openrct2/network/network.h>
-#include <openrct2/ObjectList.h>
+#include <openrct2/object/ObjectList.h>
 #include <openrct2/sprites.h>
 #include <openrct2-ui/interface/Dropdown.h>
 #include <openrct2/world/scenery.h>
