+++ src/openrct2/EditorObjectSelectionSession.h
@@ -17,7 +17,7 @@
 #pragma once
 
 #include "common.h"
-#include "object.h"
+#include "object/Object.h"
 
 #ifdef __cplusplus
 extern bool _maxObjectsWasHit;
