+++ src/openrct2/object/ObjectFactory.cpp
@@ -34,7 +34,8 @@
 #include "WallObject.h"
 #include "WaterObject.h"
 
-#include "../object.h"
+#include "../object/Object.h"
+#include "ObjectLimits.h"
 
 class ReadObjectContext : public IReadObjectContext
 {
@@ -111,7 +112,7 @@ namespace ObjectFactory
             rct_object_entry entry = fs.ReadValue<rct_object_entry>();
             result = CreateObject(entry);
 
-            utf8 objectName[9] = { 0 };
+            utf8 objectName[DAT_NAME_LENGTH + 1] = { 0 };
             object_entry_get_name_fixed(objectName, sizeof(objectName), &entry);
             log_verbose("  entry: { 0x%08X, \"%s\", 0x%08X }", entry.flags, objectName, entry.checksum);
 
@@ -144,7 +145,7 @@ namespace ObjectFactory
         Object * result = CreateObject(*entry);
         if (result != nullptr)
         {
-            utf8 objectName[9];
+            utf8 objectName[DAT_NAME_LENGTH + 1];
             object_entry_get_name_fixed(objectName, sizeof(objectName), entry);
 
             auto readContext = ReadObjectContext(objectName);
