+++ src/openrct2/object/ObjectLimits.h
@@ -1,4 +1,4 @@
-#pragma region Copyright (c) 2014-2017 OpenRCT2 Developers
+#pragma region Copyright (c) 2014-2018 OpenRCT2 Developers
 /*****************************************************************************
  * OpenRCT2, an open source clone of Roller Coaster Tycoon 2.
  *
@@ -14,20 +14,18 @@
  *****************************************************************************/
 #pragma endregion
 
-#include "../object/Object.h"
+#pragma once
 
-#ifndef _LEGACY_H
-#define _LEGACY_H
+#define MAX_RIDE_OBJECTS          128
+#define MAX_SMALL_SCENERY_OBJECTS 252
+#define MAX_LARGE_SCENERY_OBJECTS 128
+#define MAX_WALL_SCENERY_OBJECTS  128
+#define MAX_BANNER_OBJECTS         32
+#define MAX_PATH_OBJECTS           16
+#define MAX_PATH_ADDITION_OBJECTS  15
+#define MAX_SCENERY_GROUP_OBJECTS  19
+#define MAX_PARK_ENTRANCE_OBJECTS   1
+#define MAX_WATER_OBJECTS           1
+#define MAX_SCENARIO_TEXT_OBJECTS   1
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-extern const rct_object_entry RequiredSelectedObjects[7];
-extern const rct_object_entry DefaultSelectedObjects[26];
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif // _LEGACY_H
+#define DAT_NAME_LENGTH             8
