+++ src/openrct2/object/ObjectList.h
@@ -1,27 +1,17 @@
 #pragma once
 
-#include "ride/Ride.h"
-#include "world/Banner.h"
-#include "world/Entrance.h"
-#include "world/footpath.h"
-#include "world/scenery.h"
-#include "world/water.h"
+#include "../ride/Ride.h"
+#include "../world/Banner.h"
+#include "../world/Entrance.h"
+#include "../world/footpath.h"
+#include "../world/scenery.h"
+#include "../world/water.h"
+#include "ObjectLimits.h"
 
 #ifdef __cplusplus
-extern "C" {
+extern "C"
+{
 #endif
-
-#define MAX_RIDE_OBJECTS          128
-#define MAX_SMALL_SCENERY_OBJECTS 252
-#define MAX_LARGE_SCENERY_OBJECTS 128
-#define MAX_WALL_SCENERY_OBJECTS  128
-#define MAX_BANNER_OBJECTS         32
-#define MAX_PATH_OBJECTS           16
-#define MAX_PATH_ADDITION_OBJECTS  15
-#define MAX_SCENERY_GROUP_OBJECTS  19
-#define MAX_PARK_ENTRANCE_OBJECTS   1
-#define MAX_WATER_OBJECTS           1
-#define MAX_SCENARIO_TEXT_OBJECTS   1
 
 extern rct_ride_entry               *gRideEntries[MAX_RIDE_OBJECTS];
 extern rct_small_scenery_entry      *gSmallSceneryEntries[MAX_SMALL_SCENERY_OBJECTS];
