+++ src/openrct2/scenario/scenario.c
@@ -28,8 +28,8 @@
 #include "../management/NewsItem.h"
 #include "../management/Research.h"
 #include "../network/network.h"
-#include "../object.h"
-#include "../ObjectList.h"
+#include "../object/Object.h"
+#include "../object/ObjectList.h"
 #include "../OpenRCT2.h"
 #include "../ParkImporter.h"
 #include "../peep/Staff.h"
