+++ src/openrct2/world/MapGen.cpp
@@ -26,7 +26,7 @@
 #include "../Game.h"
 #include "../Imaging.h"
 #include "../localisation/string_ids.h"
-#include "../object.h"
+#include "../object/Object.h"
 #include "../platform/platform.h"
 #include "../util/Util.h"
 #include "map.h"
