+++ test/testpaint/compat.c
@@ -15,7 +15,7 @@
 #pragma endregion
 
 #include <openrct2/config/Config.h>
-#include <openrct2/object.h>
+#include <openrct2/object/Object.h>
 #include <openrct2/interface/colour.h>
 #include <openrct2/interface/viewport.h>
 #include <openrct2/ride/Ride.h>
